feat: Add sign-in text form fields File and integrate animations in sign-in view
