+++ lib/features/sign_in/presentation/widgets/sign_in_body.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:sadio_mane_store/core/common/widget/custom_icon.dart';
-import 'package:sadio_mane_store/core/common/widget/custom_text_form_field.dart';
 import 'package:sadio_mane_store/core/helpers/extensions/localization_extension.dart';
 import 'package:sadio_mane_store/core/helpers/spacer/spacer_helper.dart';
 import 'package:sadio_mane_store/core/theme/extensions/app_theme_extension.dart';
 import 'package:sadio_mane_store/features/sign_in/presentation/widgets/change_theme_and_language_row_widget.dart';
+import 'package:sadio_mane_store/features/sign_in/presentation/widgets/sign_in_text_form_fields.dart';
 
 class SignInBody extends StatelessWidget {
   const SignInBody({super.key});
@@ -57,31 +57,6 @@ class SignInBody extends StatelessWidget {
           ),
         ],
       ),
-    );
-  }
-}
-
-class SignInTextFormFields extends StatelessWidget {
-  const SignInTextFormFields({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        CustomTextFormField(hintText: context.tr.your_email),
-        verticalSpace(40),
-        CustomTextFormField(
-          hintText: context.tr.password,
-          obscureText: true,
-          suffixIcon: IconButton(
-            onPressed: () {},
-            icon: Icon(
-              Icons.visibility,
-              color: context.theme.appColors.textColor,
-            ),
-          ),
-        ),
-      ],
     );
   }
 }
